Constants and Regex
I added some constants to keep things uniform for the rest of the JS. In addition, I plan on creating error codes for the form when invalid expressions are entered. This is where the addition of regular expressions are used to create format tests for the inputs in the form.